CCC_Project_P445-P446-F24-S25/requirements.txt
alembic==1.14.1
antiorm==1.2.1
app==0.0.1
arrow==1.3.0
astroid==3.3.8
bcrypt==4.2.1
bidict==0.23.1
binaryornot==0.4.4
bleach==6.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
config==0.5.1
cookiecutter==2.1.1
create-app==0.6.2
cryptography==44.0.0
db==0.1.1
dill==0.3.9
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
hvac==2.3.0
idna==3.10
iniconfig==2.0.0
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jinja2-time==0.2.0
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
netifaces==0.11.0
nose2==0.15.1
numpy==2.2.3
packaging==24.2
pbr==6.1.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
pylint==3.3.4
PyMySQL==1.1.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-slugify==8.0.4
python-socketio==5.12.1
PyYAML==6.0.2
qiskit==1.4.1
repoze.lru==0.7
requests==2.28.1
Routes==2.5.1
rustworkx==0.16.0
scipy==1.15.2
setuptools==75.8.2
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.38
stevedore==5.4.1
symengine==0.13.0
sympy==1.13.3
text-unidecode==1.3
tomlkit==0.13.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==1.26.20
users==1.0.dev0
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1